 requirements.txt
Django>=5.1
waitress>=2.1.0
django-widget-tweaks>=1.4.12
django-otp
qrcode[pil]
pyotp
django-extensions werkzeug
pyOpenSSL
python-dateutil
pandas
reportlab
djangorestframework
psycopg2-binary
install psycopg2